requirements.txt
Django>=2.1
django-environ
django-import-export
djangorestframework
factory_boy
